заводит отдельный презентер для точки маршрута
